Add files via upload
These are authoritative articles designed to address top of the funnel interest for specific categories. These articles address a number of topics within a central concern. They are designed to show the company's knowledge and expertise. They build brand awareness and trust. Very few people convert on these pages, but they are a strong conduit to middle and bottom of the funnel concerns that do convert.